+++ care-pathway/care-pathway-service-with-loco/src/models/_entities/mod.rs
@@ -4,6 +4,7 @@ pub mod audit_logs;
 pub mod care_pathways;
 pub mod event_outbox;
 pub mod instance_events;
+pub mod instance_measures;
 pub mod instance_steps;
 pub mod instance_team;
 pub mod merge_records;